Move validation funcs out of cli module
